# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: number;
  title: string;
  docType: string;
  pageCount: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface DocumentChunk {
  id: number;
  pageNum: number;
  content: string;
}

export interface DocumentDetail {
  id: number;
  title: string;
  docType: string;
  pageCount: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  chunks: DocumentChunk[];
}

export type QueryInputSearchMode = typeof QueryInputSearchMode[keyof typeof QueryInputSearchMode];


export const QueryInputSearchMode = {
  hybrid: 'hybrid',
  keyword: 'keyword',
  vector: 'vector',
} as const;

export interface QueryInput {
  /** @minLength 1 */
  query: string;
  searchMode?: QueryInputSearchMode;
  topK?: number;
}

export interface Citation {
  documentId: number;
  documentTitle: string;
  docType: string;
  pageNum: number;
  excerpt: string;
  score: number;
  /** @nullable */
  bm25Score?: number | null;
  /** @nullable */
  vectorScore?: number | null;
}

export interface QueryResult {
  query: string;
  answer: string;
  citations: Citation[];
  searchMode: string;
  processingTimeMs: number;
  llmUsed?: boolean;
}

export interface GoldenSetEntry {
  id: number;
  query: string;
  groundTruthAnswer: string;
  sourceDocument: string;
  pageNumbers: string;
  /** @nullable */
  category?: string | null;
}

export interface EvaluationEntryResult {
  id: number;
  query: string;
  groundTruth: string;
  systemAnswer: string;
  sourceDocument: string;
  retrievalAccuracy: number;
  faithfulness: number;
  retrieved: string[];
}

export interface EvaluationReport {
  avgRetrievalAccuracy: number;
  avgFaithfulness: number;
  totalEntries: number;
  passedEntries: number;
  results: EvaluationEntryResult[];
  ranAt: string;
}

export interface SystemStatsDocumentsByType {[key: string]: number}

export interface SystemStats {
  totalDocuments: number;
  totalChunks: number;
  documentsByType: SystemStatsDocumentsByType;
  avgPageCount: number;
}