# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /documents/{id}:
    get:
      operationId: getDocument
      summary: Get a document by ID
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    DocumentChunk:
      type: object
      required: [id, pageNum, content]
      properties:
        id:
          type: integer
        pageNum:
          type: integer
        content:
          type: string
    DocumentDetail:
      type: object
      required: [id, title, docType, pageCount, createdAt, chunks]
      properties:
        id:
          type: integer
        title:
          type: string
        docType:
          type: string
        pageCount:
          type: integer
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/DocumentChunk"
...