# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentChunk } from './documentChunk';

export interface DocumentDetail {
  id: number;
  title: string;
  docType: string;
  pageCount: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  chunks: DocumentChunk[];
}